requirements.txt
telethon==1.36.0
python-dotenv==1.0.0
supabase==2.3.4
requests==2.31.0
